requirements.txt
Flask==0.9
Flask-MongoKit==0.6
Flask-OAuth==0.12
Flask-OpenID==1.1.1
Flask-RESTful==0.2.2
Flask-WTF==0.8.3
Jinja2==2.7
MarkupSafe==0.16
PyYAML==3.10
WTForms==1.0.4
Werkzeug==0.8.3
charade==1.0.3
distribute==0.6.27
flask-mongoengine==0.7.0
gunicorn==0.17.4
httplib2==0.8
mongoengine==0.8.1
mongokit==0.8.3
newrelic==1.12.0.56
oauth2==1.5.211
pyjade==2.0.2
pymongo==2.5.1
pysrt==0.5.0
python-openid==2.2.5
simplejson==3.3.0
sqlite3dbm==0.1.4
wsgiref==0.1.2
Markdown==2.3.1
waitress==0.8.8

